+++ crates/medea-macro/src/caused.rs
@@ -1,69 +1,67 @@
-use proc_macro2::{Span, TokenStream};
+use proc_macro2::TokenStream;
 use quote::quote;
 use syn::parse::{Error, Result};
 use synstructure::{BindStyle, Structure};
 
-/// Generates the actual code for `#[derive(JsCaused)]` macro.
+/// Generates the actual code for `#[derive(Caused)]` macro.
 ///
 /// # Algorithm
 ///
 /// 1. Generate body for trait method `name()` as enum variant name "as is".
-/// 2. Generate body for trait method `js_cause()`:
+/// 2. Generate body for trait method `cause()`:
 ///     - if `enum` variant contains associated error, returns this error;
-///     - if `enum` variant contains `JsCaused`, invoke its trait method;
+///     - if `enum` variant contains `Caused`, invoke its trait method;
 ///     - otherwise returns `None`.
-/// 3. Generate implementation of `JsCaused` trait for this enum with generated
+/// 3. Generate implementation of `Caused` trait for this enum with generated
 ///    methods from step 1 and 2.
 #[allow(clippy::needless_pass_by_value)]
 pub fn derive(mut s: Structure) -> Result<TokenStream> {
     let error_type = error_type(&s)?;
 
     let cause_body = s.bind_with(|_| BindStyle::Move).each_variant(|v| {
-        if let Some(js_error) =
+        if let Some(caused) = v.bindings().iter().find(|&bi| is_caused(bi)) {
+            quote!(return #caused.cause())
+        } else if let Some(error) =
             v.bindings().iter().find(|&bi| is_error(bi, &error_type))
         {
-            quote!(return Some(#js_error))
-        } else if let Some(js_caused) =
-            v.bindings().iter().find(|&bi| is_caused(bi))
-        {
-            quote!(return #js_caused.js_cause())
+            quote!(return Some(#error))
         } else {
             quote!(return None)
         }
     });
 
-    let js_caused = s.gen_impl(quote! {
+    let caused = s.gen_impl(quote! {
         #[automatically_derived]
-        gen impl JsCaused for @Self {
+        gen impl Caused for @Self {
             type Error = #error_type;
 
-            fn js_cause(self) -> Option<Self::Error> {
+            fn cause(self) -> Option<Self::Error> {
                 match self { #cause_body }
             }
         }
     });
 
-    Ok(quote!(#js_caused))
+    Ok(quote!(#caused))
 }
 
-/// Parse and returns argument of `#[js(error = "path::to::Error"))]` attribute.
-/// If no such attribute exists the defaults to `JsError`.
+/// Parse and returns argument of `#[cause(error = "path::to::Error"))]`
+/// attribute. If no such attribute exists the defaults to `JsError`.
 fn error_type(s: &synstructure::Structure) -> Result<syn::Path> {
     let mut error_type = None;
     for attr in &s.ast().attrs {
         if let Ok(meta) = attr.parse_meta() {
-            if meta.path().is_ident("js") {
+            if meta.path().is_ident("cause") {
                 if error_type.is_some() {
                     return Err(Error::new_spanned(
                         meta,
-                        "Cannot have two #[js(...)] attributes",
+                        "Cannot have two #[cause(...)] attributes",
                     ));
                 }
                 if let syn::Meta::List(list) = meta {
                     if list.nested.is_empty() {
                         return Err(Error::new_spanned(
                             list,
-                            "Expected at least one argument to #[js(...)] \
+                            "Expected at least one argument to #[cause(...)] \
                              attribute",
                         ));
                     }
@@ -90,7 +88,7 @@ fn error_type(s: &synstructure::Structure) -> Result<syn::Path> {
                             _ => {
                                 return Err(Error::new_spanned(
                                     list,
-                                    "Expected attribute like #[js(error = \
+                                    "Expected attribute like #[cause(error = \
                                      \"path::to::error\")]",
                                 ));
                             }
@@ -98,7 +96,7 @@ fn error_type(s: &synstructure::Structure) -> Result<syn::Path> {
                 } else {
                     return Err(Error::new_spanned(
                         meta,
-                        "#[js] attribute must take a list in parentheses",
+                        "#[cause] attribute must take a list in parentheses",
                     ));
                 };
             }
@@ -106,29 +104,17 @@ fn error_type(s: &synstructure::Structure) -> Result<syn::Path> {
     }
     match error_type {
         Some(path) => Ok(path),
-        None => syn::LitStr::new("JsError", Span::call_site()).parse(),
+        None => Err(Error::new_spanned(s.ast(), "Error type wasn't provided")),
     }
 }
 
-/// Checks that enum variant has `#[js(cause)]` attribute.
+/// Checks that enum variant has `#[cause]` attribute.
 fn is_caused(bi: &synstructure::BindingInfo) -> bool {
     let mut found_cause = false;
     for attr in &bi.ast().attrs {
-        if let Ok(meta) = attr.parse_meta() {
-            if meta.path().is_ident("js") {
-                if let syn::Meta::List(ref list) = meta {
-                    if let Some(syn::NestedMeta::Meta(syn::Meta::Path(
-                        ref path,
-                    ))) = list.nested.first()
-                    {
-                        if path.is_ident("cause") {
-                            if found_cause {
-                                panic!("Cannot have two cause attributes");
-                            }
-                            found_cause = true;
-                        }
-                    }
-                }
+        if let Ok(syn::Meta::Path(path)) = attr.parse_meta() {
+            if path.is_ident("cause") {
+                found_cause = true;
             }
         }
     }