+++ jason/src/peer/mod.rs
@@ -33,7 +33,7 @@ use crate::{
         MediaStreamSettings, RecvConstraints,
     },
     platform,
-    utils::JsCaused,
+    utils::Caused,
 };
 
 #[doc(inline)]
@@ -52,17 +52,17 @@ pub use self::{
 };
 
 /// Errors occurring in [`PeerConnection::update_local_stream()`] method.
-#[derive(Clone, Debug, Display, From, JsCaused)]
-#[js(error = "platform::Error")]
+#[derive(Clone, Debug, Display, From, Caused)]
+#[cause(error = "platform::Error")]
 pub enum UpdateLocalStreamError {
     /// Errors occurred when [`TracksRequest`] validation fails.
     InvalidLocalTracks(TracksRequestError),
 
     /// [`MediaManager`] failed to acquire [`local::Track`]s.
-    CouldNotGetLocalMedia(#[js(cause)] InitLocalTracksError),
+    CouldNotGetLocalMedia(#[cause] InitLocalTracksError),
 
     /// Errors occurred in [`MediaConnections::insert_local_tracks()`] method.
-    InsertLocalTracksError(#[js(cause)] InsertLocalTracksError),
+    InsertLocalTracksError(#[cause] InsertLocalTracksError),
 }
 
 /// Events emitted from a [`Sender`] or a [`Receiver`].
@@ -92,11 +92,11 @@ pub enum TrackEvent {
 
 /// Local media update errors that [`PeerConnection`] reports in
 /// [`PeerEvent::FailedLocalMedia`] messages.
-#[derive(Clone, Debug, Display, From, JsCaused)]
-#[js(error = "platform::Error")]
+#[derive(Clone, Debug, Display, From, Caused)]
+#[cause(error = "platform::Error")]
 pub enum LocalMediaError {
     /// Error occurred in [`PeerConnection::update_local_stream()`] method.
-    UpdateLocalStreamError(#[js(cause)] UpdateLocalStreamError),
+    UpdateLocalStreamError(#[cause] UpdateLocalStreamError),
 
     /// Error occurred when creating a new [`Sender`].
     ///